backend/requirements.txt
#requirements.txt
fastapi
uvicorn
qdrant-client
pydantic
python-dotenv
pydantic[email]
openai
fitz
PyMuPDF
python-multipart
textblob